Making Kotlin the main lang, finally

+++ src/main/kotlin/my/little/changelog/service/leaf/LeafService.kt
@@ -1,5 +1,6 @@
 package my.little.changelog.service.leaf
 
+import my.little.changelog.configuration.auth.CustomPrincipal
 import my.little.changelog.model.leaf.dto.service.LeafCreationDto
 import my.little.changelog.model.leaf.dto.service.LeafDeletionDto
 import my.little.changelog.model.leaf.dto.service.LeafReturnedDto
@@ -19,7 +20,8 @@ object LeafService {
 
     fun createLeaf(leaf: LeafCreationDto): Response<LeafReturnedDto> = transaction {
         val version = VersionRepo.findById(leaf.versionId)
-        VersionValidator.validateLatest(version)
+
+        VersionValidator.validateLatest(version, leaf.principal.user)
             .chain {
                 LeafValidator.validateNew(leaf)
             }
@@ -32,7 +34,7 @@ object LeafService {
     fun updateLeaf(leafUpdate: LeafUpdateDto): Response<Unit> = transaction {
         val leaf = LeafRepo.findById(leafUpdate.id)
         val newParentGroup = GroupRepo.findLatestGroupByVid(leafUpdate.parentVid)
-        VersionValidator.validateLatest(leaf.version)
+        VersionValidator.validateLatest(leaf.version, leafUpdate.principal.user)
             .chain {
                 LeafValidator.validateUpdate(leafUpdate)
             }
@@ -50,17 +52,24 @@ object LeafService {
 
     fun deleteLeaf(leafDeletionDto: LeafDeletionDto): Response<Unit> = transaction {
         val leaf = LeafRepo.findById(leafDeletionDto.id)
-        VersionValidator.validateLatest(leaf.version)
+
+        VersionValidator.validateLatest(leaf.version, leafDeletionDto.principal.user)
             .ifValid {
-                LeafRepo.delete(leaf)
+                if (leafDeletionDto.completely) {
+                    leaf.isDeleted = true
+                    LeafRepo.update(leaf)
+                } else {
+                    LeafRepo.delete(leaf)
+                }
+                Unit
             }
     }
 
-    fun changePosition(leafId: Int, changeAgainstId: Int): Response<Unit> = transaction {
+    fun changePosition(leafId: Int, changeAgainstId: Int, cp: CustomPrincipal): Response<Unit> = transaction {
         val leaf = LeafRepo.findById(leafId)
         val leafChangeAgainst = LeafRepo.findById(changeAgainstId)
-        VersionValidator.validateLatest(leaf.version)
-            .chain { VersionValidator.validateLatest(leafChangeAgainst.version) }
+        VersionValidator.validateLatest(leaf.version, cp.user)
+            .chain { VersionValidator.validateLatest(leafChangeAgainst.version, cp.user) }
             .chain {
                 ValidatorResponse
                     .ofSimple("Could not modify leaves that is not in the same group. IDs [${leaf.id.value}] [${leafChangeAgainst.id.value}]") {
